requirements.txt
flask==1.1.4
requests==2.27.1
python-dotenv==0.19.2
gunicorn==20.1.0
werkzeug==1.0.1
itsdangerous==1.1.0
click==7.1.2
jinja2==2.11.3
markupsafe==2.0.1
charset-normalizer==2.0.12
idna==3.3
urllib3==1.26.9
certifi==2021.10.8
# No color extraction dependencies needed 